Updated wolf6.aiv - Added signature statues in front of back gate's bastions. - Moar pitch near the front, for moar doom at only +1 building step! - Relocated Fbal 1 and 2 (now to be found next to Tre 1 and 2) to avoid collisions with iron mines and rocks on skirmish trail Mission 80. - Slightly adjusted killing pits to now encompass FBal 1 and 2 as well (better protection and tighter fit).
---

This update is meant to be part of a fix for #7... or actually not, but it still is part of #10, where issues of #7 and other things were fixed. With the issue already described in detail there, I will omit an explicit description of why these changes were done for now, should be obvious.